+++ server/application/user/user_routes.py
@@ -13,27 +13,12 @@ import uuid
 
 # For Development
 from settings import *
+from models import User
 
 collection = get_collection("user")
 
 
-# User Model
-class User(Document):
-    user_id = StringField(required=True)
-    name = StringField(required=True)
-    email = EmailField(max_length=50)
-    password = StringField(required=True)
-    ticket_ID = StringField(required=True)
 
-    def json(self):
-        user_dict = {
-            "user_id": self.user_id,
-            "name": self.name,
-            "email": self.email,
-            "password": self.password,
-            "ticket_ID": self.ticket_ID
-        }
-        return json.dumps(user_dict)
 
 
 # Set up a Blueprint
