+++ server/base_data/table_events.csv
@@ -1,3 +1,3 @@
 event_id,event_name,event_address,event_start_date,event_end_date,event_start_time,event_end_time
-2ab60824-b539-4a1f-ae1a-f7d94d2d55bb,Eintracht Frankfurt vs. FC Bayern München,"Mörfelder Landstraße 362, 60528 Frankfurt am Main",2020-05-30,2020-05-30,15:30:00,18:00:00
+2ab60824-b539-4a1f-ae1a-f7d94d2d55bb,Eintracht Frankfurt vs. FC Bayern München,"Mörfelder Landstraße 362, 60528 Frankfurt am Main",2020-06-01,2020-06-01,15:30:00,18:00:00
 7be626b0-c1cd-4a26-a111-60afb13ba940,World Club Dome,"Mörfelder Landstraße 362, 60528 Frankfurt am Main",2020-06-05,2020-06-07,20:00:00,18:00:00
